move importing list queries
